requirements.txt
crewai
pandas
matplotlib
faiss-cpu
sentence-transformers
chainlit
speechrecognition